Registro de Mensajeros y Solicitantes implementados
- Valores Default de formularios cambiados
- Coneccion de Aplicaciones con Mediador implementado
- Coneccion de Vistas con Gestores implementado
- Correción de DAOs para ajustarlos a nuevo modelo de BD

+++ src/datos/SolicitanteDAO.java
@@ -40,14 +40,14 @@ public class SolicitanteDAO {
         Connection conexion = ServiceLocator.getInstance().tomarConexion();
         PreparedStatement prepStmt = conexion.prepareStatement(strSQL);
         //Establece los parametros a insertar el Solicitane
-        prepStmt.setInt(1, solicitante.getK_numeroDocumento()); 
+        prepStmt.setLong(1, solicitante.getK_numeroDocumento()); 
         prepStmt.setString(2, solicitante.getK_tipoDocumento()); 
         prepStmt.setString(3, solicitante.getN_primerNombre()); 
         prepStmt.setString(4, solicitante.getN_segundoNombre()); 
         prepStmt.setString(5, solicitante.getN_primerApellido());   
         prepStmt.setString(6, solicitante.getN_segundoApellido()); 
         prepStmt.setString(7, solicitante.getN_sexo());   
-        prepStmt.setInt(8, solicitante.getQ_telefono());
+        prepStmt.setLong(8, solicitante.getQ_telefono());
         prepStmt.setString(9, solicitante.getN_correoElectronico());   
         prepStmt.setString(10, solicitante.getN_direccion());  
         //Ejecuta el registro del solicitante en la base de datos
